improve usage without import docs

+++ docs/plugins.py
@@ -39,7 +39,9 @@ def build_history():
     history = re.sub(r'#(\d+)', r'[#\1](https://github.com/samuelcolvin/python-devtools/issues/\1)', history)
     history = re.sub(r'( +)@([\w\-]+)', r'\1[@\2](https://github.com/\2)', history, flags=re.I)
     history = re.sub('@@', '@', history)
-    (THIS_DIR / '.history.md').write_text(history)
+    output_file = THIS_DIR / '.history.md'
+    if not output_file.exists() or history != output_file.read_text():
+        (THIS_DIR / '.history.md').write_text(history)
 
 
 def gen_example_html(markdown: str):
@@ -71,9 +73,11 @@ def set_version(markdown: str, page: Page) -> str:
 def remove_files(files: Files) -> Files:
     to_remove = []
     for file in files:
-        if file.src_path in {'plugins.py', 'requirements.txt'}:
+        if file.src_path == 'requirements.txt':
             to_remove.append(file)
         elif file.src_path.startswith('__pycache__/'):
+            to_remove.append(file)
+        elif file.src_path.endswith('.py'):
             to_remove.append(file)
 
     logger.debug('removing files: %s', [f.src_path for f in to_remove])
